Update packer to enable embedded module support when reading
