+++ pkg/sourcereader/embedded.go
@@ -16,7 +16,6 @@ package sourcereader
 
 import (
 	"fmt"
-	"hpc-toolkit/pkg/modulereader"
 	"io/fs"
 	"io/ioutil"
 	"os"
@@ -86,150 +85,6 @@ func copyFSToTempDir(fs BaseFS, modulePath string) (string, error) {
 	}
 	err = copyDirFromModules(fs, modulePath, tmpDir)
 	return tmpDir, err
-}
-
-// GetModuleInfo gets modulereader.ModuleInfo for the given kind from the embedded source
-func (r EmbeddedSourceReader) GetModuleInfo(modPath string, kind string) (modulereader.ModuleInfo, error) {
-	if !IsEmbeddedPath(modPath) {
-		return modulereader.ModuleInfo{}, fmt.Errorf("Source is not valid: %s", modPath)
-	}
-
-	reader := modulereader.Factory(kind)
-	mi, err := reader.GetInfo(modPath)
-	mi.RequiredApis = defaultAPIList(modPath)
-	return mi, err
-}
-
-func defaultAPIList(source string) []string {
-	// API lists at
-	// https://console.cloud.google.com/apis/dashboard and
-	// https://console.cloud.google.com/apis/library
-	staticAPIMap := map[string][]string{
-		"community/modules/compute/SchedMD-slurm-on-gcp-partition": {
-			"compute.googleapis.com",
-		},
-		"community/modules/compute/htcondor-execute-point": {
-			"compute.googleapis.com",
-		},
-		"community/modules/compute/pbspro-execution": {
-			"compute.googleapis.com",
-			"storage.googleapis.com",
-		},
-		"community/modules/compute/schedmd-slurm-gcp-v5-partition": {
-			"compute.googleapis.com",
-		},
-		"community/modules/database/slurm-cloudsql-federation": {
-			"bigqueryconnection.googleapis.com",
-			"sqladmin.googleapis.com",
-		},
-		"community/modules/file-system/DDN-EXAScaler": {
-			"compute.googleapis.com",
-			"deploymentmanager.googleapis.com",
-			"iam.googleapis.com",
-			"runtimeconfig.googleapis.com",
-		},
-		"community/modules/file-system/Intel-DAOS": {
-			"compute.googleapis.com",
-			"iam.googleapis.com",
-			"secretmanager.googleapis.com",
-		},
-		"community/modules/file-system/nfs-server": {
-			"compute.googleapis.com",
-		},
-		"community/modules/project/new-project": {
-			"admin.googleapis.com",
-			"cloudresourcemanager.googleapis.com",
-			"cloudbilling.googleapis.com",
-			"iam.googleapis.com",
-		},
-		"community/modules/project/service-account": {
-			"iam.googleapis.com",
-		},
-		"community/modules/project/service-enablement": {
-			"serviceusage.googleapis.com",
-		},
-		"community/modules/scheduler/SchedMD-slurm-on-gcp-controller": {
-			"compute.googleapis.com",
-		},
-		"community/modules/scheduler/SchedMD-slurm-on-gcp-login-node": {
-			"compute.googleapis.com",
-		},
-		"modules/scheduler/batch-job-template": {
-			"batch.googleapis.com",
-			"compute.googleapis.com",
-		},
-		"modules/scheduler/batch-login-node": {
-			"batch.googleapis.com",
-			"compute.googleapis.com",
-			"storage.googleapis.com",
-		},
-		"community/modules/scheduler/htcondor-configure": {
-			"iam.googleapis.com",
-			"secretmanager.googleapis.com",
-		},
-		"community/modules/scheduler/pbspro-client": {
-			"compute.googleapis.com",
-			"storage.googleapis.com",
-		},
-		"community/modules/scheduler/pbspro-server": {
-			"compute.googleapis.com",
-			"storage.googleapis.com",
-		},
-		"community/modules/scheduler/schedmd-slurm-gcp-v5-controller": {
-			"compute.googleapis.com",
-			"iam.googleapis.com",
-			"pubsub.googleapis.com",
-			"secretmanager.googleapis.com",
-		},
-		"community/modules/scheduler/schedmd-slurm-gcp-v5-hybrid": {
-			"compute.googleapis.com",
-			"pubsub.googleapis.com",
-		},
-		"community/modules/scheduler/schedmd-slurm-gcp-v5-login": {
-			"compute.googleapis.com",
-		},
-		"community/modules/scripts/htcondor-install": {},
-		"community/modules/scripts/omnia-install":    {},
-		"community/modules/scripts/pbspro-preinstall": {
-			"iam.googleapis.com",
-			"storage.googleapis.com",
-		},
-		"community/modules/scripts/pbspro-install": {},
-		"community/modules/scripts/pbspro-qmgr":    {},
-		"community/modules/scripts/spack-install":  {},
-		"community/modules/scripts/wait-for-startup": {
-			"compute.googleapis.com",
-		},
-		"modules/compute/vm-instance": {
-			"compute.googleapis.com",
-		},
-		"modules/file-system/filestore": {
-			"file.googleapis.com",
-		},
-		"modules/file-system/pre-existing-network-storage": {},
-		"modules/monitoring/dashboard": {
-			"stackdriver.googleapis.com",
-		},
-		"modules/network/pre-existing-vpc": {
-			"compute.googleapis.com",
-		},
-		"modules/network/vpc": {
-			"compute.googleapis.com",
-		},
-		"modules/packer/custom-image": {
-			"compute.googleapis.com",
-			"storage.googleapis.com",
-		},
-		"modules/scripts/startup-script": {
-			"storage.googleapis.com",
-		},
-	}
-
-	requiredAPIs, found := staticAPIMap[source]
-	if !found {
-		return []string{}
-	}
-	return requiredAPIs
 }
 
 // GetModule copies the embedded source to a provided destination (the deployment directory)
