+++ pkg/sourcereader/sourcereader.go
@@ -16,7 +16,6 @@ package sourcereader
 
 import (
 	"hpc-toolkit/pkg/deploymentio"
-	"hpc-toolkit/pkg/modulereader"
 	"log"
 	"strings"
 )
@@ -32,7 +31,6 @@ type SourceReader interface {
 	// GetModuleInfo would leverage modulereader.GetInfo for the given kind.
 	// GetModuleInfo would operate over the source without creating a local copy.
 	// This would be very dependent on the kind of module.
-	GetModuleInfo(modPath string, kind string) (modulereader.ModuleInfo, error)
 
 	// GetModule copies the source to a provided local destination (the deployment directory).
 	GetModule(modPath string, copyPath string) error
