+++ pkg/sourcereader/sourcereader_test.go
@@ -15,7 +15,6 @@
 package sourcereader
 
 import (
-	"embed"
 	"fmt"
 	"io/ioutil"
 	"log"
@@ -42,8 +41,6 @@ const (
 // Setup GoCheck
 type MySuite struct{}
 
-//go:embed modules
-var testModuleFS embed.FS
 var _ = Suite(&MySuite{})
 
 func Test(t *testing.T) {
